Add Links to Each Memory Allocator
Add links to each memory allocator, with official websites being used as the link targets when possible; GitHub repositories of the memory allocators used as fallback link targets when no official website could be found. This change will allow users of this repository to quickly and easily get more information on and access each memory allocator outside of this repository.