hardened_malloc has MTE support now
Pixel 8 and Pixel 8 Pro were launched with MTE support and it's deployed
in the latest release of GrapheneOS as an opt-in option in user facing
security settings rather than hidden in developer options. It will be
enabled by default for the base OS including system apps along with user
installed apps without their own native libraries. It will be opt-in for
apps with their own native libraries along with executables separate
from their app processes.